Move functions into the plugin
