+++ packages/demo/package.json
@@ -2,7 +2,6 @@
   "name": "demo",
   "version": "1.0.0",
   "dependencies": {
-    "flat-cache": "^3.0.4",
     "gatsby": "^3.7.0-next.0",
     "gatsby-plugin-gatsby-cloud": "^2.5.0",
     "gatsby-plugin-image": "^1.5.0",
