+++ packages/gatsby-plugin-support-chat/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@loadable/component": "^5.15.0",
+    "flat-cache": "^3.0.4",
     "gatsby": "^3.7.0-next.0"
   }
 }
