+++ packages/gatsby-plugin-support-chat/src/api/gatsby-plugin-support-chat/send-message.ts
